Fix #1109 - test_readmes.py::test_simple_remote fails when insufficient funds in wallet for remote network
