+++ .github/workflows/libcheck.yml
@@ -50,7 +50,7 @@ jobs:
           FACTORY_DEPLOYER_PRIVATE_KEY: "0xc594c6e5def4bab63ac29eed19a134c130388f74f019bc74b8f4389df2837a58"
         run: |
           mkcodes --github --output tests/generated-readmes/test_{name}.{ext} READMEs
-          py.test tests/test_readmes.py
+          pytest tests/readmes/test_readmes.py
       - name: Slack notify via webhook
         uses: up9cloud/action-notify@master
         if: cancelled() == false
