+++ tests/integration/remote/test_polygon.py
@@ -12,47 +12,24 @@ from ocean_lib.models.datatoken import Datatoken
 from ocean_lib.ocean.ocean import Ocean
 from ocean_lib.web3_internal.utils import connect_to_network
 
-from .util import get_wallets, random_chars, set_aggressive_gas_fees
+from . import util
 
 
 @pytest.mark.integration
-def test_ocean_tx__create_url_asset(tmp_path):
-    """On Polygon, do the Ocean txs for create_url_asset()."""
+def test_ocean_tx__create_data_nft(tmp_path):
+    """On Polygon, do a simple Ocean tx: create_data_nft"""
     # setup
     connect_to_network("polygon")
-    set_aggressive_gas_fees()
+    util.set_aggressive_gas_fees()
 
-    config = _remote_config_polygon(tmp_path)
+    config = util.remote_config_polygon(tmp_path)
     ocean = Ocean(config)
+
     accounts.clear()
-    (alice_wallet, _) = get_wallets(ocean)
-
-    # Alice call create_url_asset
-    # avoid "replacement transaction underpriced" error: make each tx diff't
-    name = random_chars()
-    url = "https://arweave.net/qctEbPb3CjvU8LmV3G_mynX74eCxo1domFQIlOBH1xU"
-    try:  # it can get away with "insufficient funds" errors, but not others
-        print("Call create_url_asset(), and wait for it to complete...")
-        (_, datatoken, _) = ocean.assets.create_url_asset(
-            name, url, alice_wallet, wait_for_aqua=False
-        )
-        assert isinstance(datatoken, Datatoken)
-
-    except ValueError as error:
-        if "insufficient funds" in str(error):
-            warnings.warn(UserWarning("Warning: Insufficient Polygon MATIC"))
-            return
-        raise error
-
-    print("Success")
+    (alice_wallet, _) = util.get_wallets()
+    
+    # Do a simple-as-possible test that uses ocean stack, while accounting for gotchas
+    util.do_ocean_tx_and_handle_gotchas(ocean, alice_wallet)
 
 
-def _remote_config_polygon(tmp_path):
-    config = {
-        "NETWORK_NAME": "polygon",
-        "METADATA_CACHE_URI": "https://v4.aquarius.oceanprotocol.com",
-        "PROVIDER_URL": "https://v4.provider.polygon.oceanprotocol.com",
-        "DOWNLOADS_PATH": "consume-downloads",
-    }
 
-    return config
