+++ tests/integration/remote/util.py
@@ -6,11 +6,33 @@ import os
 import random
 import string
 import time
+import warnings
 
 from brownie.network import accounts, chain, priority_fee
 
+def remote_config_mumbai(tmp_path):
+    config = {
+        "NETWORK_NAME": "mumbai",
+        "METADATA_CACHE_URI": "https://v4.aquarius.oceanprotocol.com",
+        "PROVIDER_URL": "https://v4.provider.mumbai.oceanprotocol.com",
+        "DOWNLOADS_PATH": "consume-downloads",
+    }
 
-def get_wallets(ocean):
+    return config
+
+
+def remote_config_polygon(tmp_path):
+    config = {
+        "NETWORK_NAME": "polygon",
+        "METADATA_CACHE_URI": "https://v4.aquarius.oceanprotocol.com",
+        "PROVIDER_URL": "https://v4.provider.polygon.oceanprotocol.com",
+        "DOWNLOADS_PATH": "consume-downloads",
+    }
+
+    return config
+
+
+def get_wallets():
     alice_private_key = os.getenv("REMOTE_TEST_PRIVATE_KEY1")
     bob_private_key = os.getenv("REMOTE_TEST_PRIVATE_KEY2")
 
@@ -28,12 +50,60 @@ def get_wallets(ocean):
     return (alice_wallet, bob_wallet)
 
 
+def set_aggressive_gas_fees():
+    # Polygon & Mumbai uses EIP-1559. So, dynamically determine priority fee
+    priority_fee(chain.priority_fee)
+
+
+def do_nonocean_tx_and_handle_gotchas(ocean, alice_wallet, bob_wallet):
+    """Call wallet.transfer(), but handle several gotchas for this test use case:
+    - if the test has to repeat, there are nonce errors. Avoid via unique
+    - if there are insufficient funds, since they're hard to replace
+      automatically in remote testnets, then just skip
+   """
+    # Simplest possible tx: Alice send Bob some fake MATIC
+    bob_eth_before = accounts.at(bob_wallet.address).balance()
+
+    normalized_unixtime = time.time() / 1e9
+    amt_send = 1e-8 * (random.random() + normalized_unixtime)
+
+    print("Do a send-Ether tx...")
+    try:  # it can get away with "insufficient funds" errors, but not others
+        alice_wallet.transfer(bob_wallet.address, f"{amt_send:.15f} ether")
+    except ValueError as error:
+        if "insufficient funds" in str(error):
+            warnings.warn(UserWarning("Warning: Insufficient test MATIC"))
+            return
+        raise (error)
+
+    bob_eth_after = accounts.at(bob_wallet.address).balance()
+    assert bob_eth_after > bob_eth_before
+
+
+def do_ocean_tx_and_handle_gotchas(ocean, alice_wallet):
+    """Call create_data_nft(), but handle several gotchas for this test use case:
+    - if the test has to repeat, there are nonce errors. Avoid via unique
+    - if there are insufficient funds, since they're hard to replace
+      automatically in remote testnets, then just skip
+   """
+    # Alice publish data NFT
+    # avoid "replacement transaction underpriced" error: make each tx diff't
+    symbol = random_chars()
+    try:  # it can get away with "insufficient funds" errors, but not others
+        print("Call create_data_nft(), and wait for it to complete...")
+        data_nft = ocean.create_data_nft(symbol, symbol, alice_wallet)
+
+    except ValueError as error:
+        if "insufficient funds" in str(error):
+            warnings.warn(UserWarning("Warning: Insufficient test MATIC"))
+            return
+        raise (error)
+
+    assert data_nft.symbol() == symbol
+    print("Success")
+
+
 def random_chars() -> str:
     cand_chars = string.ascii_uppercase + string.digits
     s = "".join(random.choices(cand_chars, k=8)) + str(time.time())
     return s
-
-
-def set_aggressive_gas_fees():
-    # Polygon & Mumbai uses EIP-1559. So, dynamically determine priority fee
-    priority_fee(chain.priority_fee)
